backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
attrs==26.1.0
certifi==2026.2.25
chardet==7.2.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
Faker==25.0.0
fastapi==0.111.0
fastapi-cli==0.0.24
frozenlist==1.8.0
greenlet==3.3.2
groq==0.13.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
orjson==3.11.7
passlib==1.7.4
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.3
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.19.2
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.22
python-telegram-bot==21.3
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.1.0
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.37.2
twilio==9.0.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
